+++ src/Nest/Domain/Stats/IndexDocStats.cs
@@ -6,13 +6,13 @@ using Newtonsoft.Json;
 
 namespace Nest
 {
-    public class IndexDocStats
-    {
-        [JsonProperty("num_docs")]
-        public int NumberOfDocs { get; set; }
-        [JsonProperty("max_docs")]
-        public int MaximumDocs { get; set; }
-        [JsonProperty("deleted_docs")]
-        public int DeletedDocs { get; set; }
-    }
+	public class IndexDocStats
+	{
+		[JsonProperty("num_docs")]
+		public long NumberOfDocs { get; set; }
+		[JsonProperty("max_docs")]
+		public long MaximumDocs { get; set; }
+		[JsonProperty("deleted_docs")]
+		public long DeletedDocs { get; set; }
+	}
 }
